+++ CHANGELOG.rst
@@ -1,5 +1,7 @@
 3.14:
 
+* `#81 <https://github.com/xitrum-framework/xitrum/issues/81>`_
+  Reload classes in development mode
 * `#393 <https://github.com/xitrum-framework/xitrum/issues/393>`_
   For development mode, set Netty's ResourceLeakDetector level to ADVANCED
 * `#395 <https://github.com/xitrum-framework/xitrum/issues/395>`_